Rename quota/mysql to quota/mysqlqm
