+++ testonly/integration/registry.go
@@ -16,7 +16,7 @@ package integration
 
 import (
 	"github.com/google/trillian/extension"
-	mysqlq "github.com/google/trillian/quota/mysql"
+	"github.com/google/trillian/quota/mysqlqm"
 	"github.com/google/trillian/storage/mysql"
 )
 
@@ -32,6 +32,6 @@ func NewRegistryForTests(testID string) (extension.Registry, error) {
 		AdminStorage: mysql.NewAdminStorage(db),
 		LogStorage:   mysql.NewLogStorage(db, nil),
 		MapStorage:   mysql.NewMapStorage(db),
-		QuotaManager: &mysqlq.QuotaManager{DB: db, MaxUnsequencedRows: mysqlq.DefaultMaxUnsequenced},
+		QuotaManager: &mysqlqm.QuotaManager{DB: db, MaxUnsequencedRows: mysqlqm.DefaultMaxUnsequenced},
 	}, nil
 }
